Restructure of Survey
